# docs/swagger.yaml
basePath: /
definitions:
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_handlers.LeaveApprovalResponseDoc:
    properties:
      leave:
        $ref: '#/definitions/internal_handlers.LeaveResponseDoc'
      staffing_risk:
        $ref: '#/definitions/internal_handlers.StaffingRiskResponseDoc'
    type: object
  internal_handlers.LeaveResponseDoc:
    properties:
      approved_at:
        example: "2026-05-23T10:00:00Z"
        type: string
      approved_by:
        example: 1
        type: integer
      document_path:
        example: uploads/leaves/leave_10_123.pdf
        type: string
      end_date:
        example: "2026-06-03T00:00:00Z"
        type: string
      id:
        example: 5
        type: integer
      leave_type:
        example: sakit
        type: string
      reason:
        example: Demam tinggi
        type: string
      start_date:
        example: "2026-06-01T00:00:00Z"
        type: string
      status:
        example: approved
        type: string
      user_id:
        example: 10
        type: integer
    type: object
...
  internal_handlers.StaffingRiskResponseDoc:
    properties:
      estimated_available_staff:
        example: 7
        type: integer
      message:
        example: Risiko Understaffed jika disetujui
        type: string
      minimum_staffing:
        example: 8
        type: integer
      warning:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend HRIS dan operasional frozen-food retail dengan JWT auth, geofenced
    attendance + selfie, leave approval with staffing risk, payroll, PPE/asset compliance,
    dan executive reports.
  title: HRIS Monitoring API
  version: "1.0"
paths:
...
  /leaves:
    get:
      description: Return leave history for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.LeaveResponseDoc'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List current user's leaves
      tags:
      - Leaves
    post:
      consumes:
      - multipart/form-data
      description: Submit leave or permit request with optional supporting document
        upload.
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: formData
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: formData
        name: end_date
        required: true
        type: string
      - description: Leave type
        in: formData
        name: leave_type
        required: true
        type: string
      - description: Reason
        in: formData
        name: reason
        type: string
      - description: Supporting document
        in: formData
        name: document
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.LeaveResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Apply leave
      tags:
      - Leaves
  /leaves/{id}:
    get:
      description: Retrieve one leave request by identifier.
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.LeaveResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get leave by ID
      tags:
      - Leaves
  /leaves/approve/{id}:
    post:
      description: Approve a pending leave request and return staffing risk evaluation.
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.LeaveApprovalResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Approve leave
      tags:
      - Leaves
  /leaves/reject/{id}:
    post:
      description: Reject a pending leave request without staffing-risk payload.
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.LeaveApprovalResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reject leave
      tags:
      - Leaves
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"